Cms (#75)
* 맵퍼s 리뷰.xml삭제

* 제품목록 상세 장바구니

* 1020 0.0.6 ok

* temp commit

* temp1541

* 결제 구현(수정필요)

* temp1702

* ui개선

* 리스트, 검색, css 수정

+++ marketDEMO/src/main/java/kr/co/team3/dto/product/ProductDTO.java
@@ -42,7 +42,7 @@ public class ProductDTO {
     @Column(name = "p_view_count")
     private int pViewCount;
     @Column(name = "star_avg")
-    private int starAvg;
+    private Double starAvg;
     @Column(name = "s_company_name")
     private String sCompanyName;
 }
